Split generation process
