+++ src/asyncapi_python_codegen/__init__.py
@@ -1,6 +1,6 @@
 """AsyncAPI Python Code Generator."""
 
-from .generator import CodeGenerator
+from .generators import CodeGenerator
 from .parser import extract_all_operations, load_document_info
 from .cli import app
