+++ src/asyncapi_python_codegen/generators/__init__.py
@@ -1,0 +1,5 @@
+"""Code generators for AsyncAPI specifications."""
+
+from .main import CodeGenerator
+
+__all__ = ["CodeGenerator"]
